requirements.txt
weaviate-client==4.5.4